requirements.txt
streamlit
chromadb
pandas
numpy
torch
transformers
Pillow
opencv-python
